refactor: форма редактирования данных пользователя вынесена в виджет ProfileEditInfoForm
