+++ src/entities/form/types.d.ts
@@ -2,6 +2,13 @@ import { Input, Textarea } from "shared/ui";
 
 declare global {
   export type TFormFields = Input | Textarea;
+
+  export type TFormPartials = {
+    form: Form;
+    fields: {
+      [k: string]: TFormFields;
+    };
+  };
 }
 
 export {};
