+++ src/widgets/profile_edit_info-form/index.ts
@@ -1,0 +1,1 @@
+export * from "./profile_edit_info-form";
